Extract dialog rendering into wallets_screen/dialogs.rs module
Moved 4 dialog rendering functions and 10+ helper methods (~1151 lines)
from wallets_screen/mod.rs into a new dialogs.rs module. Includes
send, receive, fund platform, and private key dialogs plus their
state structs. mod.rs reduced from 3824 to 2673 lines.

Task: 3.2a
